poller/requirements.txt
paramiko
pyyaml
requests
mysql-connector-python
